requirements.txt

streamlit>=1.34.0
pandas>=2.1.0
numpy>=1.26.0
gspread>=6.0.0
google-auth>=2.30.0
gspread-dataframe>=3.3.0
xlsxwriter>=3.2.0
matplotlib>=3.8.0
